Split apart DB Util functions
